Final version for Kontrol pausability proofs (#9530)
* Update tests to native symbolic `bytes` and `bytes[]`

* adding lemmas

* `run-kontrol.sh`: add more sensible parameters

* Change `startPrank` by `prank`

* Replace `mockCall` workaround with `vm.mockCall`

* Make `bytes` length symbolic

* `KontrolUtils`: remove symbolic workarounds

* `run-kontrol.sh`: add `prove_proveWithdrawalTransaction_paused`

* `forge fmt`

* `versions.json`: bump Kontrol from `0.1.127` to `0.1.156`

* Remove `ASSUME` comments

* ci: run kontrol on develop and allow manual dispatch

* ci: rename parameter

* `OptimismPortalKontrol`: add remaining ranges for `_withdrawalProof`

* Add forge-like UX to `run-kontrol.sh`

* `pausability-lemmas.k`: clean file

* general tests, further lemmas, summary claim

* Address shellcheck failures

* Change `pausability-lemmas.k` to `pausability-lemmas.md`

* `versions.json`: bump `kontrol` from `0.1.156` to `0.1.178`

* `OptimismPortalKontrol`: update `kontrol` natspec to version 0.1.178

* `pausability-lemmas.md`: remove unused `Lemmas` header

* Update packages/contracts-bedrock/test/kontrol/pausability-lemmas.md


* Update packages/contracts-bedrock/test/kontrol/pausability-lemmas.md


* Update packages/contracts-bedrock/test/kontrol/pausability-lemmas.md


* Update packages/contracts-bedrock/test/kontrol/pausability-lemmas.md


* `pausability-lemmas.md`: fix spelling typo

* `pausability-lemmas.md`: fix typo `bytearrays` to `byte arrays`

* `run-kontrol.sh`: correctly format temorarily unexecuted lemmas

* `common.sh`: execute `copy_to_docker` only when in docker mode

* `make-summary-deployment.sh`: add argument check before parsing

* Reflect `kontrol summary` change to `kontrol load-state-diff`

From version `0.1.162`, `kontrol summary` has been renamed to `kontrol
load-state-diff`.
The reason of this renaming is that `kontrol summary` will be
used by kontrol's compositional symbolic execution.
Also, changing the name to `load-state-diff` makes more explicit what the
command does.
Related PR: https://github.com/runtimeverification/kontrol/pull/374

* `pausability-lemmas.md`: remove upstreamed lemmas

* fix: writing typos

* lemma text pass

* paragraph about summary maintainability

* README.md: include latest changes

* pausability-lemmas.md: markdown link typo

* KontrolUtils: add documentation comment

* pausability-lemmas.md: fix markdown typo vol2

* pausability-lemmas.md: fix markdown typo vol3

* Add specialized usage functions

* `README.md`: remove `bash` in `make-summary-deployment.sh` description

* `README.md`: complete `Add New Proofs` section

* versions.json: bump kontrol from 0.1.178 to 0.1.196

* run-kontrol.sh: add `--xml-test-report` flag

* .gitignore: add `kontrol_prove_report.xml`

* foundry.toml: set `ast = true` in `kprove` profile

* config.yml: set correct path for kontrol `store_artifacts`

* config.yml: add `store_test_results` step to `kontrol-tests` job

* package.json: execute `run-kontrol.sh` with `script` option

We run `run-kontrol.sh` with the `script` option to avoid executing all proofs

* run-kontrol.sh: remove proof with 10 elements in favor of 0 and 1

The longer the `_withdrawalProof` array the longer the execution time
Adding the lengths 0 and 1 fits within the max cpus and won't take as long to run

* chore: typos and formatting

* ci: fix kontrol trigger

* README.md: minor typo

---------

+++ packages/contracts-bedrock/test/kontrol/scripts/common.sh
@@ -2,10 +2,31 @@
 # Common functions and variables for run-kontrol.sh and make-summary-deployment.sh
 
 notif() { echo "== $0: $*" >&2 ; }
-usage() {
-  echo "Usage: $0 [-h|--help] [container|local|dev]" 1>&2
-  echo "Options:" 1>&2
+
+# usage function for the run-kontrol.sh script
+usage_run_kontrol() {
+  echo "Usage: $0 [-h|--help] [container|local|dev] [script|tests]" 1>&2
+  echo "" 1>&2
   echo "  -h, --help         Display this help message." 1>&2
+  echo "" 1>&2
+  echo "Execution modes:"
+  echo "  container          Run in docker container. Reproduce CI execution. (Default)" 1>&2
+  echo "  local              Run locally, enforces registered versions.json version for better reproducibility. (Recommended)" 1>&2
+  echo "  dev                Run locally, does NOT enforce registered version. (Useful for developing with new versions and features)" 1>&2
+  echo "" 1>&2
+  echo "Tests executed:"
+  echo "  script             Execute the tests recorded in run-kontrol.sh" 1>&2
+  echo "  tests              Execute the tests provided as arguments" 1>&2
+  exit 0
+}
+
+# usage function for the make-summary-deployment.sh script
+usage_make_summary() {
+  echo "Usage: $0 [-h|--help] [container|local|dev]" 1>&2
+  echo "" 1>&2
+  echo "  -h, --help         Display this help message." 1>&2
+  echo "" 1>&2
+  echo "Execution modes:"
   echo "  container          Run in docker container. Reproduce CI execution. (Default)" 1>&2
   echo "  local              Run locally, enforces registered versions.json version for better reproducibility. (Recommended)" 1>&2
   echo "  dev                Run locally, does NOT enforce registered version. (Useful for developing with new versions and features)" 1>&2
@@ -20,25 +41,84 @@ export CONTAINER_NAME=kontrol-tests
 KONTROLRC=$(jq -r .kontrol < "$WORKSPACE_DIR/../../versions.json")
 export KONTROL_RELEASE=$KONTROLRC
 export LOCAL=false
+export SCRIPT_TESTS=false
+SCRIPT_OPTION=false
+export CUSTOM_TESTS=0 # Store the position where custom tests start, interpreting 0 as no tests
+CUSTOM_OPTION=0
+export RUN_KONTROL=false # true if any functions are called from run-kontrol.sh
+
+# General usage function, which discerns from which script is being called and displays the appropriate message
+usage() {
+  if [ "$RUN_KONTROL" = "true" ]; then
+    usage_run_kontrol
+  else
+    usage_make_summary
+  fi
+}
+
 
 # Argument Parsing
+# The logic behind argument parsing is the following (in order):
+# - Execution mode argument: container (or empty), local, dev
+# - Tests arguments (first if execution mode empty): script, specific test names
 parse_args() {
-  if [ $# -gt 1 ]; then
+  if [ $# -eq 0 ]; then
+    export LOCAL=false
+    export SCRIPT_TESTS=false
+    export CUSTOM_TESTS=0
+  # `script` argument caps the total possible arguments to its position
+  elif { [ $# -gt 1 ] && [ "$1" == "script" ]; } || { [ $# -gt 2 ] && [ "$2" == "script" ]; }; then
     usage
-  elif [ $# -eq 0 ] || [ "$1" == "container" ]; then
+  elif [ $# -eq 1 ]; then
+    SCRIPT_OPTION=false
+    CUSTOM_OPTION=0
+    parse_first_arg "$1"
+  elif [ $# -eq 2 ] && [ "$2" == "script" ]; then
+    if [ "$1" != "container" ] && [ "$1" != "local" ] && [ "$1" != "dev" ]; then
+      notif "Invalid first argument. Must be \`container\`, \`local\` or \`dev\`"
+      exit 1
+    fi
+    SCRIPT_OPTION=true
+    CUSTOM_OPTION=0
+    parse_first_arg "$1"
+  else
+    SCRIPT_OPTION=false
+    CUSTOM_OPTION=2
+    parse_first_arg "$1"
+  fi
+}
+
+# Parse the first argument passed to `run-kontrol.sh`
+parse_first_arg() {
+  if [ "$1" == "container" ]; then
     notif "Running in docker container (DEFAULT)"
     export LOCAL=false
+    export SCRIPT_TESTS=$SCRIPT_OPTION
+    export CUSTOM_TESTS=$CUSTOM_OPTION
   elif [ "$1" == "-h" ] || [ "$1" == "--help" ]; then
     usage
   elif [ "$1" == "local" ]; then
     notif "Running with LOCAL install, .kontrolrc CI version ENFORCED"
+    export SCRIPT_TESTS=$SCRIPT_OPTION
+    export CUSTOM_TESTS=$CUSTOM_OPTION
     check_kontrol_version
   elif [ "$1" == "dev" ]; then
     notif "Running with LOCAL install, IGNORING .kontrolrc version"
     export LOCAL=true
+    export SCRIPT_TESTS=$SCRIPT_OPTION
+    export CUSTOM_TESTS=$CUSTOM_OPTION
     pushd "$WORKSPACE_DIR" > /dev/null || exit
+  elif [ "$1" == "script" ]; then
+    notif "Running in docker container (DEFAULT)"
+    export LOCAL=false
+    NEGATED_SCRIPT_TESTS=$([[ "${SCRIPT_OPTION}" == "true" ]] && echo false || echo true)
+    export SCRIPT_TESTS=$NEGATED_SCRIPT_TESTS
+    export CUSTOM_TESTS=$CUSTOM_OPTION
   else
-    usage
+    notif "Running in docker container (DEFAULT)"
+    export LOCAL=false
+    export SCRIPT_TESTS=$SCRIPT_OPTION
+    export CUSTOM_TESTS=1 # Store the position where custom tests start
   fi
 }
 
@@ -87,13 +167,15 @@ copy_to_docker() {
   #   Error response from daemon: invalid symlink "/home/user/workspace/node_modules/@typescript-eslint/eslint-plugin" -> "../../../../node_modules/.pnpm/@typescript-eslint+eslint-plugin@6.19.1_@typescript-eslint+parser@6.19.1_eslint@8.56.0_typescript@5.3.3/node_modules/@typescript-eslint/eslint-plugin"
   # Even though we use a bind mount, we still need to copy the files to the
   # container because we are running Docker on a remote host.
-  TMP_DIR=$(mktemp -d)
-  cp -r "$WORKSPACE_DIR/." "$TMP_DIR"
-  rm -rf "$TMP_DIR/node_modules"
-  docker cp --follow-link "$TMP_DIR/." $CONTAINER_NAME:/home/user/workspace
-  rm -rf "$TMP_DIR"
+  if [ "$LOCAL" == false ]; then
+    TMP_DIR=$(mktemp -d)
+    cp -r "$WORKSPACE_DIR/." "$TMP_DIR"
+    rm -rf "$TMP_DIR/node_modules"
+    docker cp --follow-link "$TMP_DIR/." $CONTAINER_NAME:/home/user/workspace
+    rm -rf "$TMP_DIR"
 
-  docker exec --user root "$CONTAINER_NAME" chown -R user:user /home/user
+    docker exec --user root "$CONTAINER_NAME" chown -R user:user /home/user
+  fi
 }
 
 clean_docker(){
